requirements.txt
numpy>=1.23.0,<2.0.0
tensorflow>=2.16.0,<2.18.0
pandas==1.3.4
scikit-learn==1.0.1
# streamlit==1.2.0
streamlit==1.37.0
streamlit_option_menu==0.3.2
pillow==8.4.0
